+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # unreleased
-    - BREAKING: 
+    - BREAKING: You can no longer add routes after the app has been dispatched